Move shortcut handling to a tool
